Revert "Merge branch 'feat/crossplatform-mk' of github.com:Jake-Carter/msdk into fix/add_uart_int_ob"
This reverts commit c50d7bfaad9d4fe29168b83b7cdd09bbfe09ba73, reversing
changes made to b2966c5e9c312c0c54b5132df1876dcd04f8d10d.

+++ Examples/MAX32655/TFT_Demo/src/utils.c
@@ -45,14 +45,14 @@ unsigned int utils_get_time_ms(void)
 {
     uint32_t sec, ssec;
     double subsec;
-    uint32_t ms;
+    unsigned int ms;
 
     MXC_RTC_GetSubSeconds(&ssec);
-    subsec = (double)ssec / (double)(4096.0);
+    subsec = ssec / 4096.0;
 
     MXC_RTC_GetSeconds(&sec);
 
-    ms = (sec * 1000) + (int)(subsec * 1000);
+    ms = (sec * 1000) + (subsec * 1000);
 
     return ms;
 }
